+++ unisim_lib/unisim/component/cxx/processor/arm/vmsav8/cpu.hh
@@ -315,7 +315,7 @@ struct CPU
   enum branch_type_t { B_JMP = 0, B_CALL, B_RET };
   void BranchTo( uint64_t addr, branch_type_t branch_type ) { next_insn_addr = addr; }
   
-  void CallSupervisor( uint16_t imm );
+  void CallSupervisor( uint32_t imm );
   
   //=====================================================================
   //=                       Memory access methods                       =
