+++ unisim_simulators/cxx/s3pac/main.cc
@@ -745,7 +745,7 @@ struct Processor
   void WaitForInterrupt() { not_implemented(); }
   void SWI( uint32_t imm ) { not_implemented(); }
   void BKPT( uint32_t imm ) { not_implemented(); }
-  void CallSupervisor( uint16_t imm ) { not_implemented(); }
+  void CallSupervisor( uint32_t imm ) { not_implemented(); }
   bool IntegerZeroDivide( BOOL const& condition ) { return false; }
   
   //   =====================================================================
